% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonframe-package.R
\docType{package}
\name{jsonframe-package}
\alias{jsonframe}
\alias{jsonframe-package}
\title{jsonframe: Process Data In Browser}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Maciej Banas \email{banasmaciek@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
